The ERA5 annual seasonal maximum Fire Weather Index and Monthly Summary Record of Fire Weather Index for the region of New South Wales from 1978 to 2019

The ERA5 annual seasonal maximum Fire Weather Index and Monthly Summary Record of Fire Weather Index for the region of New South Wales from 1978 to 2019

Usage

data

Format

'data'

A data frame with 43 rows and 3 columns

year

the year of the value

FWI

The annual seasonal maximum Fire Weather Index

MSR

Monthly Summary Record of Fire Weather Index

Source

https://climexp.knmi.nl/bushfires_timeseries.cgi